Make savefile access its own class
Let's add a class called Storage to handle reading and writing to the savefile.
